+++ src/api/models/Answers.ts
@@ -2,7 +2,5 @@
 /* istanbul ignore file */
 /* tslint:disable */
 /* eslint-disable */
-
 import type { Answer } from './Answer';
-
 export type Answers = Array<Answer>;