+++ src/api/models/Domains.ts
@@ -2,7 +2,5 @@
 /* istanbul ignore file */
 /* tslint:disable */
 /* eslint-disable */
-
 import type { Domain } from './Domain';
-
 export type Domains = Array<Domain>;