+++ src/api/models/Group.ts
@@ -2,10 +2,8 @@
 /* istanbul ignore file */
 /* tslint:disable */
 /* eslint-disable */
-
 import type { Domains } from './Domains';
 import type { Questions } from './Questions';
-
 export type Group = {
     id: number;
     greeting: string;