+++ src/api/models/Question.ts
@@ -2,9 +2,7 @@
 /* istanbul ignore file */
 /* tslint:disable */
 /* eslint-disable */
-
 import type { Answers } from './Answers';
-
 export type Question = {
     text: string;
     answers: Answers;