+++ src/api/models/Questions.ts
@@ -2,7 +2,5 @@
 /* istanbul ignore file */
 /* tslint:disable */
 /* eslint-disable */
-
 import type { Question } from './Question';
-
 export type Questions = Array<Question>;