+++ src/api/services/AuthService.ts
@@ -3,13 +3,10 @@
 /* tslint:disable */
 /* eslint-disable */
 import type { TokenData } from '../models/TokenData';
-
 import type { CancelablePromise } from '../core/CancelablePromise';
 import { OpenAPI } from '../core/OpenAPI';
 import { request as __request } from '../core/request';
-
 export class AuthService {
-
     /**
      * Get token data
      * @returns TokenData GetTokenResponse
@@ -32,5 +29,4 @@ export class AuthService {
             },
         });
     }
-
 }