+++ src/api/services/InfoService.ts
@@ -3,13 +3,10 @@
 /* tslint:disable */
 /* eslint-disable */
 import type { Healthcheck } from '../models/Healthcheck';
-
 import type { CancelablePromise } from '../core/CancelablePromise';
 import { OpenAPI } from '../core/OpenAPI';
 import { request as __request } from '../core/request';
-
 export class InfoService {
-
     /**
      * Healthcheck
      * @returns Healthcheck HealthcheckResponse
@@ -21,5 +18,4 @@ export class InfoService {
             url: '/healthcheck',
         });
     }
-
 }