+++ docker/entrypoint.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env sh
+
+envsubst < /usr/share/nginx/html/env.src.js > /usr/share/nginx/html/env.js
+
+nginx -g "daemon off;"