+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 2024-09-19 unreleased
   - Use fixed codegen version.
+  - Support for confugurable from env API url.
 
 2024-01-10 v0.1.6
   - Temporary fix domain.